third_party/WebKit/LayoutTests/fast/dom/Window/getMatchedCSSRules-with-invalid-pseudo-elements-expected.txt

CONSOLE WARNING: line 1: document.getMatchedCSSRules() is deprecated and will be removed in M63, around December 2017. See https://www.chromestatus.com/features/4606972603138048 for more details.
getMatchedCSSRules with invalid Pseudo Elements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getMatchedCSSRules(document.querySelector('p'), '')[0].cssText is "p { color: green; }"
PASS getMatchedCSSRules(document.querySelector('p'), undefined)[0].cssText is "p { color: green; }"
PASS getMatchedCSSRules(document.querySelector('p'), null)[0].cssText is "p { color: green; }"
PASS getMatchedCSSRules(document.querySelector('p'))[0].cssText is "p { color: green; }"
PASS getMatchedCSSRules(document.querySelector('p'), 'before') is null
PASS getMatchedCSSRules(document.querySelector('p'), 'foo') is null
PASS getMatchedCSSRules(document.querySelector('div'), 'first-letter')[0].cssText is "div::first-letter { color: red; }"
PASS getMatchedCSSRules(document.querySelector('div'), ':first-letter')[0].cssText is "div::first-letter { color: red; }"
PASS getMatchedCSSRules(document.querySelector('div'), '::first-letter')[0].cssText is "div::first-letter { color: red; }"
PASS successfullyParsed is true

TEST COMPLETE
Test for https://bugs.webkit.org/show_bug.cgi?id=83298

First letter of this sentence should be red.
